last few incompatibilities python 2 to 3. Code should be now compatible with both python 2 and 3

+++ classes/atmosphere.py
@@ -389,7 +389,7 @@ class atmosphere(object):
 
                 logging.info('Running inteprolation with  %i threads' % self.nthreads)
 
-                nruns = self.nlayers/self.nthreads
+                nruns = int(self.nlayers/self.nthreads)
                 remainder = self.nlayers % self.nthreads
 
                 for round_idx in range(nruns+1):
